Separated model classes into individual files. Added IsDirtyInfo and AssociateProperties for debugging. Updated CustomTabCollection to propagate item prop changes.
