+++ Views/CustomTabUserControl.xaml
@@ -40,7 +40,8 @@
             </ListView.ItemTemplate>
         </ListView>
 
-        <TextBlock Text="{Binding IsDirty, StringFormat=IsDirty {0}, NotifyOnSourceUpdated=True}" 
+        <!-- DEBUG: Bind to IsDirtyInfo -->
+        <TextBlock Text="{Binding IsDirtyInfo, StringFormat=IsDirty {0}, NotifyOnSourceUpdated=True}" 
                    FontSize="20" FontWeight="Bold" 
                    HorizontalAlignment="Center" />
